# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "API Extender REST API",
    "version": "v1"
  },
  "paths": {
...
    "/api/regex-tool/find-matches": {
      "post": {
        "operationId": "findMatches",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "regex": {
                    "description": "Regular expression",
                    "type": "string"
                  },
                  "text": {
                    "description": "Text to search",
                    "type": "string"
                  }
                },
                "required": [
                  "regex",
                  "text"
                ],
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Regex matches found"
          }
        },
        "summary": "Find regex matches in text",
        "tags": [
          "Regex Tool"
        ]
      }
    },
...
  },
...
}